reroll classes add user class
