+++ src/test/java/org/example/UserInfo/UserInfo.java
@@ -1,0 +1,4 @@
+package org.example.UserInfo;
+
+public class UserInfo {
+}
